python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Arnav Jindal <arnav.jindal7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio-tools = "^1.54.2"
grpcio = "^1.54.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
